refactor(macro_argument_binding): split into submodules
594 lines was getting unwieldy. Split the rule into a four-file
module:

- `macro_argument_binding.rs` — module root: lint declaration,
  registration, the pre-expansion `EarlyLintPass` impl,
  `PENDING_VIOLATIONS` queue.
- `macro_argument_binding/config.rs` — `Mode`, `Config`, the
  `MacroArgumentBinding` state struct, the curated built-in lists,
  and the `parse_path_list` / `merge_with_builtins` helpers.
- `macro_argument_binding/triviality.rs` — argument splitter,
  `looks_like_expression` heuristic, and the trivial-expression
  token-stream walker (`take_*`-style per
  `IMPLEMENTATION_CONVENTIONS.md`).
- `macro_argument_binding/late.rs` — `LateLintPass` impl,
  `EnclosingHirFinder`, and the `emit` helper.

No behaviour change — every public API of the rule's module root
(`register_lint`, `register_pass`, `MACRO_ARGUMENT_BINDING`) keeps
its previous signature. Submodules use `pub(super)` /
`pub(crate)` visibility to scope helpers tightly.

Also takes the opportunity to simplify `merge_with_builtins` to
a single `chain().collect()` instead of `extend`, and adds a
short module doc-comment to each new file.
